// apps/server/src/generated/prisma/models/AiLeadQueueConfig.ts
export type AiLeadQueueConfigModel = runtime.Types.Result.DefaultSelection<Prisma.$AiLeadQueueConfigPayload>
// ...
export type $AiLeadQueueConfigPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "AiLeadQueueConfig"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    configKey: string
    workerConcurrency: number
    priorityStrategy: string
    updatedById: string | null
    updatedByName: string | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["aiLeadQueueConfig"]>
  composites: {}
}